Update and patch failing tests
