+++ subprojects/model-core/src/testFixtures/groovy/org/gradle/internal/reflect/validation/ValidationMessageChecker.groovy
@@ -169,7 +169,7 @@ trait ValidationMessageChecker {
         config.description("has @Input annotation used on property of type '${config.propertyType}'")
             .reason("A property of type '${config.propertyType}' annotated with @Input cannot determine how to interpret the file")
             .solution("Annotate with @InputFile for regular files")
-            .solution("Annotate with @InputDirectory for directories")
+            .solution("Annotate with @InputFiles for collections of files")
             .solution("If you want to track the path, return File.absolutePath as a String and keep @Input")
             .render()
     }
