+++ subprojects/plugin-development/src/integTest/groovy/org/gradle/plugin/devel/tasks/AbstractPluginValidationIntegrationSpec.groovy
@@ -444,200 +444,6 @@ abstract class AbstractPluginValidationIntegrationSpec extends AbstractIntegrati
         RegularFileProperty.name        | "getProject().getObjects().fileProperty().fileValue(new java.io.File(\"input.txt\"))"
     }
 
-    @ValidationTestFor([
-        ValidationProblemId.MISSING_NORMALIZATION_ANNOTATION,
-        ValidationProblemId.INCORRECT_USE_OF_INPUT_ANNOTATION
-    ])
-    def "detects problems with file inputs"() {
-        file("input.txt").text = "input"
-        file("input").createDir()
-
-        javaTaskSource << """
-            import org.gradle.api.*;
-            import org.gradle.api.file.*;
-            import org.gradle.api.tasks.*;
-            import java.util.Set;
-            import java.util.Collections;
-            import java.io.File;
-            import java.nio.file.Path;
-
-            @CacheableTask
-            public abstract class MyTask extends DefaultTask {
-                @javax.inject.Inject
-                org.gradle.api.internal.file.FileResolver fileResolver;
-
-                @InputDirectory
-                @Optional
-                public File getInputDirectory() {
-                    return new File("input");
-                }
-
-                @InputFile
-                public File getInputFile() {
-                    return new File("input.txt");
-                }
-
-                @InputFiles
-                public Set<File> getInputFiles() {
-                    return Collections.emptySet();
-                }
-
-                @Input
-                public File getFile() {
-                    return new File("some-file");
-                }
-
-                @Input
-                public FileCollection getFileCollection() {
-                    return getProject().files();
-                }
-
-                @Input
-                public Path getFilePath() {
-                    return new File("some-file").toPath();
-                }
-
-                @Input
-                public FileTree getFileTree() {
-                    return getProject().files().getAsFileTree();
-                }
-
-                @TaskAction
-                public void doStuff() { }
-            }
-        """
-
-        expect:
-        executer.withArgument("-Dorg.gradle.internal.max.validation.errors=10")
-        assertValidationFailsWith([
-            error(incorrectUseOfInputAnnotation { type('MyTask').property('file').propertyType('File') }, 'validation_problems', 'incorrect_use_of_input_annotation'),
-            error(incorrectUseOfInputAnnotation { type('MyTask').property('fileCollection').propertyType('FileCollection') }, 'validation_problems', 'incorrect_use_of_input_annotation'),
-            error(incorrectUseOfInputAnnotation { type('MyTask').property('filePath').propertyType('Path') }, 'validation_problems', 'incorrect_use_of_input_annotation'),
-            error(incorrectUseOfInputAnnotation { type('MyTask').property('fileTree').propertyType('FileTree') }, 'validation_problems', 'incorrect_use_of_input_annotation'),
-            error(missingNormalizationStrategy { type('MyTask').property('inputDirectory').annotatedWith('InputDirectory') }, 'validation_problems', 'missing_normalization_annotation'),
-            error(missingNormalizationStrategy { type('MyTask').property('inputFile').annotatedWith('InputFile') }, 'validation_problems', 'missing_normalization_annotation'),
-            error(missingNormalizationStrategy { type('MyTask').property('inputFiles').annotatedWith('InputFiles') }, 'validation_problems', 'missing_normalization_annotation'),
-        ])
-    }
-
-    @ValidationTestFor(
-        ValidationProblemId.MISSING_ANNOTATION
-    )
-    def "detects problems on nested collections"() {
-        javaTaskSource << """
-            import org.gradle.api.*;
-            import org.gradle.api.provider.*;
-            import org.gradle.api.tasks.*;
-            import org.gradle.work.*;
-            import java.util.*;
-            import java.io.File;
-
-            @DisableCachingByDefault(because = "test task")
-            public class MyTask extends DefaultTask {
-                @Nested
-                public Options getOptions() {
-                    return new Options();
-                }
-
-                @Nested
-                public List<Options> getOptionsList() {
-                    return Arrays.asList(new Options());
-                }
-
-                @Nested
-                public Iterable<Options> getIterableOptions() {
-                    return Arrays.asList(new Options());
-                }
-
-                @Nested
-                public Iterable<Iterable<Options>> getDoubleIterableOptions() {
-                    return Arrays.asList(Arrays.asList(new Options()));
-                }
-
-                @Nested
-                public Map<String, Options> getMappedOptions() {
-                    return Collections.singletonMap("alma", new Options());
-                }
-
-                @Nested
-                public Iterable<Map<String, Iterable<Options>>> getIterableMappedOptions() {
-                    return Arrays.asList(Collections.singletonMap("alma", Arrays.asList(new Options())));
-                }
-
-                @Nested
-                public Provider<Options> getProvidedOptions() {
-                    return getProject().getObjects().property(Options.class).convention(new Options());
-                }
-
-                @Nested
-                public Iterable<NamedBean> getNamedIterable() {
-                    return Arrays.asList(new NamedBean());
-                }
-
-                @Nested
-                public AnnotatedList getAnnotatedList() {
-                    return new AnnotatedList();
-                }
-
-                public static class Options {
-                    @Input
-                    public String getGood() {
-                        return "good";
-                    }
-
-                    public String getNotAnnotated() {
-                        return null;
-                    }
-                }
-
-                public static class NamedBean implements Named {
-                    @Input
-                    public String getGood() {
-                        return "good";
-                    }
-
-                    public String getNotAnnotated() {
-                        return null;
-                    }
-
-                    @Internal
-                    public String getName() {
-                        return "tibor";
-                    }
-                }
-
-                // Does not validate the type parameter of extended collection
-                // because it has annotated properties
-                public static class AnnotatedList extends ArrayList<Options> {
-                    public AnnotatedList() {
-                        add(new Options());
-                    }
-
-                    @Input
-                    public String getGood() {
-                        return "good";
-                    }
-                }
-
-                @TaskAction
-                public void doStuff() { }
-            }
-        """
-
-        expect:
-        executer.withArgument("-Dorg.gradle.internal.max.validation.errors=10")
-        assertValidationFailsWith([
-            error(missingAnnotationMessage { type('MyTask').property("doubleIterableOptions${iterableSymbol}${iterableSymbol}.notAnnotated").missingInputOrOutput() }, 'validation_problems', 'missing_annotation'),
-            error(missingAnnotationMessage { type('MyTask').property("iterableMappedOptions${iterableSymbol}${getKeySymbolFor("alma")}${iterableSymbol}.notAnnotated").missingInputOrOutput() }, 'validation_problems', 'missing_annotation'),
-            error(missingAnnotationMessage { type('MyTask').property("iterableOptions${iterableSymbol}.notAnnotated").missingInputOrOutput() }, 'validation_problems', 'missing_annotation'),
-            error(missingAnnotationMessage { type('MyTask').property("mappedOptions${getKeySymbolFor("alma")}.notAnnotated").missingInputOrOutput() }, 'validation_problems', 'missing_annotation'),
-            error(missingAnnotationMessage { type('MyTask').property("namedIterable${getNameSymbolFor("tibor")}.notAnnotated").missingInputOrOutput() }, 'validation_problems', 'missing_annotation'),
-            error(missingAnnotationMessage { type('MyTask').property("options.notAnnotated").missingInputOrOutput() }, 'validation_problems', 'missing_annotation'),
-            error(missingAnnotationMessage { type('MyTask').property("optionsList${iterableSymbol}.notAnnotated").missingInputOrOutput() }, 'validation_problems', 'missing_annotation'),
-            error(missingAnnotationMessage { type('MyTask').property("providedOptions.notAnnotated").missingInputOrOutput() }, 'validation_problems', 'missing_annotation'),
-        ])
-    }
-
     @ValidationTestFor(
         ValidationProblemId.PRIVATE_GETTER_MUST_NOT_BE_ANNOTATED
     )
